requirements.txt
# Core
openai>=1.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# ML
scikit-learn>=1.5.0
xgboost>=2.0.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.4.0

# RAG / Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Data generation
Faker>=25.0.0

# Visualization
plotly>=5.22.0
matplotlib>=3.9.0

# HTTP
httpx>=0.27.0
jsonschema>=4.19.0
cachetools>=5.4.0
pytest>=7.4.0
pytest-asyncio>=0.22.0
prometheus-client>=0.16.0
redis>=5.0.0
